requirements.txt
accelerate==1.12.0
adapters==1.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
airllm==2.11.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
apache-tvm-ffi==0.1.8.post2
astor==0.8.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
backoff==2.2.1
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
blake3==1.0.8
blingfire==0.1.8
blinker==1.9.0
brotli==1.2.0
cached-property==2.0.1
cachetools==7.0.1
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
comm==0.2.3
compressed-tensors==0.13.0
contourpy==1.3.2
cryptography==46.0.5
cuda-bindings==13.1.1
cuda-pathfinder==1.3.5
cuda-python==13.1.1
cupy-cuda12x==14.0.1
cycler==0.12.1
datasets==4.6.1
debugpy==1.8.20
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.133.0
fastapi-cli==0.0.24
fastapi-cloud-cli==0.14.0
fastar==0.8.0
ffmpy==1.0.0
filelock==3.24.3
filetype==1.2.0
flashinfer-python==0.6.3
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gguf==0.17.1
gradio==6.8.0
gradio_client==2.2.0
groovy==0.1.2
grpcio==1.78.1
grpcio-reflection==1.78.1
h11==0.16.0
h2==4.3.0
haystack-ai==2.25.1
haystack-experimental==0.19.0
hf-xet==1.3.1
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ijson==3.5.0
importlib_metadata==8.7.1
interegular==0.3.3
ipykernel==7.2.0
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lark==1.2.2
lazy_imports==1.2.0
litellm==1.75.0
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.2
magika==0.6.3
markdown-it-py==4.0.0
markdownify==1.2.2
markitdown==0.1.5
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.9.1
model-hosting-container-standards==0.1.13
more-itertools==10.8.0
mosestokenizer==1.2.1
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.18
neo4j==5.28.2
nest_asyncio==1.6.0
networkx==2.8.8
ninja==1.13.0
nltk==3.9.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.4.0
nvidia-cutlass-dsl-libs-base==4.4.0
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime==1.23.2
openai==2.24.0
openai-harmony==0.0.8
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openfile==0.0.7
optimum==2.1.0
orjson==3.11.7
outlines_core==0.2.11
packaging==26.0
pandas==2.3.3
parso==0.8.6
partial-json-parser==0.2.1.1.post7
pexpect==4.9.0
pickleshare==0.7.5
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.2
portalocker==3.2.0
posthog==7.9.4
prettytable==3.17.0
prometheus_client==0.24.1
prometheus-fastapi-instrumentator==7.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==23.0.1
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.17.0
qwen-vl-utils==0.0.14
ray==2.54.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.2.3
sentencepiece==0.2.1
sentry-sdk==2.53.0
setproctitle==1.3.7
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
skops==0.13.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.2.0
stack_data==0.6.3
starlette==0.52.1
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
toolwrapper==2.1.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
triton==3.5.1
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uctools==1.3.0
umap-learn==0.5.5
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
vllm==0.16.0
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
wtpsplit==2.1.7
xgrammar==0.1.29
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0

